félkész

+++ Modulzaro/Jatekos.cs
@@ -11,8 +11,8 @@ namespace Modulzaro
         public string nev  { get; set; }
         public string elso { get; set; }
         public string utolso { get; set; }
-        public int suly { get; set; }
-        public int magassag { get; set; }
+        public double suly { get; set; }
+        public double magassag { get; set; }
 
         public Jatekos(string sor)
         {
@@ -20,8 +20,8 @@ namespace Modulzaro
             nev = sorok[0];
             elso = sorok[1];
             utolso = sorok[2];
-            suly = int.Parse(sorok[3]);
-            magassag = int.Parse(sorok[4]);
+            suly = double.Parse(sorok[3]);
+            magassag = double.Parse(sorok[4]);
         }
     }
 }
